+++ Api/Request/Contract/IAddProductDimensionsContract.cs
@@ -1,6 +1,6 @@
 ﻿namespace Api.Request.Contract
 {
-    public interface IAddProductDimensionsContract : IProductDimensionsContract<int?, int>
+    public interface IAddProductDimensionsContract : IProductDimensionsContract<decimal?, decimal>
     {
     }
 }
